requirements.txt
fastdtw==0.3.4
h5py==3.6.0
jiwer==2.3.0
kaldiio==2.17.2
librosa==0.9.1
neptune-client==0.16.1
parallel-wavegan==0.5.4
pysptk==0.1.20
pyworld==0.3.0
PyYAML==6.0
resampy==0.2.2
Resemblyzer==0.1.1.dev0
tornado==6.2
tqdm
transformers==4.18.0
